requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
datasets==4.0.0
dill==0.3.8
filelock==3.19.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.3.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
kiwisolver==1.4.9
matplotlib==3.10.6
multidict==6.6.4
multiprocess==0.70.16
numpy==2.3.2
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
pyarrow==21.0.0
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
six==1.17.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
